[#601] lib: move app specific lib files to lib folder in apps
